Change webpacker/webpack configuration to dynamically load publicPath based on meta header
